-refactor: enum class 따로 선언해서 관리
